Move JaiaToggle into its own directory with its tests.
